Edited based on https://docs.microsoft.com/en-us/windows/uwp/launch-resume/launch-settings-app#ms-settings-uri-scheme-reference
Note that a number of confirmed-working entries did not exist in that page.  I have kept them.

Everything has been tested and works.  However, if it's listed in that document and doesn't work for me I've kept it.  Things I had previously noted, which Microsoft doesn't, and which don't work for me, have been deleted.